Register checkpoints on Lap 1
I keep a register of the checkpoints on lap 1 to use that info to perform better no next laps. I divided the board in cells 32 x 32 and keep the cell used info in two 32 bit ints, this way of register the cell use is faster than the linear check i was doing on the previous commit.